move 'sparseCheckpoints' one level up and use it in the wallet layer to avoid overloading the database
